refactor: move validator command tests to a separate file
